Bump image dependency + version

+++ src/color_cut_quantizer.rs
@@ -1,5 +1,6 @@
-use crate::{filter::Filter, swatch::Swatch};
 use std::collections::{BinaryHeap, HashMap};
+
+use crate::{filter::Filter, swatch::Swatch};
 
 const QUANTIZE_WORD_WIDTH: u32 = 5;
 const QUANTIZE_WORD_MAX: u8 = (1 << QUANTIZE_WORD_WIDTH) - 1;
@@ -46,11 +47,12 @@ where
         // begin by generating a histogram of quantized pixel values
         let mut hist = HashMap::new();
         for pixel in self.pixels.iter() {
-            let pixel = pixel.map(|channel| modify_width(channel, 8, QUANTIZE_WORD_WIDTH) as u8);
+            let pixel = pixel.map(|channel| modify_width(channel, 8, QUANTIZE_WORD_WIDTH));
             *hist.entry(pixel).or_insert(0) += 1;
         }
 
-        // convert the histogram into a collection of (color, count) tuples, filtering out unwanted colors
+        // convert the histogram into a collection of (color, count) tuples, filtering out unwanted
+        // colors
         let hist_len = hist.len();
         let mut colors = hist
             .into_iter()
@@ -63,15 +65,19 @@ where
             })
             .collect::<Vec<_>>();
 
-        // the colors have to be ordered at this point, so order them by combining their channels into a single integer
-        // where the red channel is the most signifcant and the blue the least
+        // the colors have to be ordered at this point, so order them by combining their channels
+        // into a single integer where the red channel is the most signifcant and the blue
+        // the least
         colors.sort_by_key(|(pixel, _)| {
             let (r, g, b) = pixel_to_rgb(pixel);
-            ((r as u32) << (QUANTIZE_WORD_WIDTH + QUANTIZE_WORD_WIDTH)) | ((g as u32) << QUANTIZE_WORD_WIDTH) | b as u32
+            ((r as u32) << (QUANTIZE_WORD_WIDTH + QUANTIZE_WORD_WIDTH))
+                | ((g as u32) << QUANTIZE_WORD_WIDTH)
+                | b as u32
         });
 
         if hist_len <= self.max_colors {
-            // there are less colors than requested, no need for further processing; just return each color as a swatch
+            // there are less colors than requested, no need for further processing; just return
+            // each color as a swatch
             colors
                 .into_iter()
                 .map(|(pixel, count)| Swatch::new(pixel_to_rgb(&pixel), count))
@@ -82,8 +88,9 @@ where
     }
 
     fn quantize_pixels(self, mut colors: Vec<(P, u32)>) -> Vec<Swatch> {
-        // create a priority queue of Vboxes with the first one containing all the given colors. Vbox comparison is
-        // based on their volume, reversed, so the queue always pops the largest Vbox by volume first
+        // create a priority queue of Vboxes with the first one containing all the given colors.
+        // Vbox comparison is based on their volume, reversed, so the queue always pops the
+        // largest Vbox by volume first
 
         let mut pq = BinaryHeap::with_capacity(self.max_colors);
         pq.push(Vbox::new(&mut colors));
@@ -107,7 +114,9 @@ where
 
     fn should_ignore_color(&self, rgb: (u8, u8, u8)) -> bool {
         let hsl = crate::rgb_to_hsl(rgb);
-        self.filters.iter().any(|filter| !filter.is_allowed(rgb, hsl))
+        self.filters
+            .iter()
+            .any(|filter| !filter.is_allowed(rgb, hsl))
     }
 
     fn split_boxes(&self, pq: &mut BinaryHeap<Vbox<'_, P>>) {
@@ -124,7 +133,8 @@ where
                 }
             }
 
-            // if the queue is empty or the largest one cannot be split, there are no more Vboxes to split
+            // if the queue is empty or the largest one cannot be split, there are no more Vboxes to
+            // split
             return;
         }
     }
@@ -207,25 +217,27 @@ where
     fn sort_colors_by_longest_dimension(&mut self) {
         let longest_dimension = self.get_longest_dimension();
 
-        self.colors.sort_by(|(lhs, _), (rhs, _)| match longest_dimension {
-            Component::Red => pixel_to_rgb(lhs).0.cmp(&pixel_to_rgb(rhs).0),
-            Component::Green => pixel_to_rgb(lhs).1.cmp(&pixel_to_rgb(rhs).1),
-            Component::Blue => pixel_to_rgb(lhs).2.cmp(&pixel_to_rgb(rhs).2),
-        });
+        self.colors
+            .sort_by(|(lhs, _), (rhs, _)| match longest_dimension {
+                Component::Red => pixel_to_rgb(lhs).0.cmp(&pixel_to_rgb(rhs).0),
+                Component::Green => pixel_to_rgb(lhs).1.cmp(&pixel_to_rgb(rhs).1),
+                Component::Blue => pixel_to_rgb(lhs).2.cmp(&pixel_to_rgb(rhs).2),
+            });
     }
 
     fn find_split_point(&mut self) -> usize {
         let midpoint = self.population / 2;
         let mut pop = 0;
 
-        // keep a total sum of all the color populations and return the first one that crosses the midpoint. if no such
-        // color is found, return the first index to still split the Vbox in two
+        // keep a total sum of all the color populations and return the first one that crosses the
+        // midpoint. if no such color is found, return the first index to still split the
+        // Vbox in two
         for (i, (_, count)) in self.colors.iter().enumerate() {
             pop += count;
 
             if pop >= midpoint {
-                // in case the first color (index 0) already crosses the midpoint, return the color after it in order to
-                // always split the Vbox in two
+                // in case the first color (index 0) already crosses the midpoint, return the color
+                // after it in order to always split the Vbox in two
                 return i.max(1);
             }
         }
@@ -252,21 +264,21 @@ where
     }
 
     fn get_average_color(&self) -> Swatch {
-        // calculate the sum of all the color populations, as well as weighted sums of each color channel based on the
-        // color populations
-        let (pop, red_sum, green_sum, blue_sum) =
-            self.colors
-                .iter()
-                .fold((0, 0, 0, 0), |(pop, red_sum, green_sum, blue_sum), (pixel, count)| {
-                    let (r, g, b) = pixel_to_rgb(pixel);
+        // calculate the sum of all the color populations, as well as weighted sums of each color
+        // channel based on the color populations
+        let (pop, red_sum, green_sum, blue_sum) = self.colors.iter().fold(
+            (0, 0, 0, 0),
+            |(pop, red_sum, green_sum, blue_sum), (pixel, count)| {
+                let (r, g, b) = pixel_to_rgb(pixel);
 
-                    (
-                        pop + count,
-                        red_sum + r as u32 * count,
-                        green_sum + g as u32 * count,
-                        blue_sum + b as u32 * count,
-                    )
-                });
+                (
+                    pop + count,
+                    red_sum + r as u32 * count,
+                    green_sum + g as u32 * count,
+                    blue_sum + b as u32 * count,
+                )
+            },
+        );
 
         // calculate the means of the channel weighted sums...
         let red_mean = red_sum as f32 / pop as f32;
